% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSurvival.R
\name{checkInput}
\alias{checkInput}
\title{\code{checkInput}}
\usage{
checkInput(input)
}
\arguments{
\item{input}{A list containing input values.}
}
\value{
A logical value indicating whether the input is valid.
}
\description{
\code{checkInput}
}
