% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountSVsDf.R
\name{CountSVsDf}
\alias{CountSVsDf}
\title{`CountSVsDf`}
\usage{
CountSVsDf(
  ncol,
  nrow,
  disease_genes_names,
  sample_names,
  genes_with_svs_in_sample,
  vcf,
  ids_col
)
}
\arguments{
\item{ncol}{name column}

\item{nrow}{name row}

\item{disease_genes_names}{disease name}

\item{sample_names}{sample name}

\item{genes_with_svs_in_sample}{gene}

\item{vcf}{file}
}
\value{
sample_disease_gene_df
}
\description{
`CountSVsDf`
}
