% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesCountTable.R
\name{genesCountTable}
\alias{genesCountTable}
\title{genesCountTable}
\usage{
genesCountTable(vcf, metadata, input, gene_ids_table)
}
\arguments{
\item{vcf}{file}

\item{metadata}{file}

\item{input}{input}

\item{gene_ids_table}{table}
}
\value{
count_table
}
\description{
genesCountTable
}
