% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSurvival.R
\name{get_disease_gene_list}
\alias{get_disease_gene_list}
\title{`get_disease_gene_list` parses an existing .txt
file containing all known target genes associated
with the selected disease.}
\usage{
get_disease_gene_list(disease_gene, input_disease)
}
\arguments{
\item{input_disease:}{input disease selection}
}
\value{
genes_list
}
\description{
`get_disease_gene_list` parses an existing .txt
file containing all known target genes associated
with the selected disease.
}
