% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSurvival.R
\name{\%+=\%}
\alias{\%+=\%}
\title{implementation of += operator
https://stackoverflow.com/questions/5738831/}
\usage{
e1 \%+=\% e2
}
\description{
implementation of += operator
https://stackoverflow.com/questions/5738831/
}
