% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_col_names.R
\name{map_col_names}
\alias{map_col_names}
\title{map_col_names}
\usage{
map_col_names(input, cov_list)
}
\arguments{
\item{input}{input}

\item{cov_list}{list}
}
\value{
cov_list
}
\description{
map_col_names
}
