% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSurvival.R
\name{Remove_NA}
\alias{Remove_NA}
\title{\code{Remove_NA} Removes rows from a data frame where the specified
time column has missing values or values equal to "NA".}
\usage{
Remove_NA(df, time_col)
}
\arguments{
\item{df:}{A data frame containing the dataset.}

\item{time_col:}{A character string specifying the name
of the time column to check.}
}
\value{
A modified data frame with rows removed
based on the specified time column.
}
\description{
\code{Remove_NA} Removes rows from a data frame where the specified
time column has missing values or values equal to "NA".
}
