% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveNAs.R
\name{RemoveNAs}
\alias{RemoveNAs}
\title{`RemoveNAs` removes all rows with Na
entries in the time factor column.}
\usage{
RemoveNAs(df, time_col)
}
\arguments{
\item{df:}{metadata data frame}

\item{time_col:}{time factor input}
}
\value{
df
}
\description{
`RemoveNAs` removes all rows with Na
entries in the time factor column.
}
