% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSurvival.R
\name{round_df}
\alias{round_df}
\title{\code{round_df} Round Numeric Columns in a Data Frame}
\usage{
round_df(x, digits)
}
\arguments{
\item{x}{A data frame to be modified.}

\item{digits}{The number of decimal places to round the
numeric columns to.}
}
\description{
\code{round_df} Round Numeric Columns in a Data Frame
}
