% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSurvival.R
\name{VariantSurvival}
\alias{VariantSurvival}
\title{VariantSurvival}
\usage{
VariantSurvival(vcf_file, metadata_file, demo = FALSE)
}
\arguments{
\item{vcf_file}{path to the vcf file containing the Structural variant data}

\item{metadata_file}{path to the txt file containing the samples metadata}

\item{demo}{true or false}
}
\description{
VariantSurvival
}
