% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSurvival.R
\name{CountSVsDf}
\alias{CountSVsDf}
\title{\code{CountSVsDf}}
\usage{
CountSVsDf(
  ncol,
  nrow,
  disease_genes_names,
  sample_names,
  genes_with_svs_in_sample,
  vcf,
  ids_col
)
}
\description{
\code{CountSVsDf}
}
