% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_df.R
\name{round_df}
\alias{round_df}
\title{round_df
round all numeric variables}
\usage{
round_df(x, digits)
}
\arguments{
\item{x}{data frame}

\item{digits}{number of digits to round}
}
\value{
x
}
\description{
round_df
round all numeric variables
}
