% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSurvival.R
\name{VariantSurvival}
\alias{VariantSurvival}
\title{VariantSurvival}
\usage{
VariantSurvival(vcffile, metadatafile)
}
\arguments{
\item{vcffile}{path to the vcf file containing the Structural variant data}

\item{metadatafile}{path to the txt file containing the samples metadata}
}
\description{
VariantSurvival
}
